fix: cleanup bootstrap functions and invocations
