+++ functions/source/BootstrapAit/crhelper/resource_helper.pyi
