--- functions/source/BootstrapEfs/crhelper-2.0.11.dist-info/INSTALLER
@@ -1,1 +1,0 @@
-pip
