--- functions/source/BootstrapEfs/crhelper-2.0.11.dist-info/NOTICE
@@ -1,2 +1,0 @@
-Custom Resource Helper
-Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
