--- functions/source/BootstrapEfs/crhelper-2.0.11.dist-info/top_level.txt
@@ -1,1 +1,0 @@
-crhelper
