+++ functions/source/ConfigParser/config.json
@@ -1,0 +1,97 @@
+{
+  "Questions": [
+    {
+      "Label": "UseRoute53",
+      "Value": "false"
+    },
+    {
+      "Label": "DeployIam",
+      "Value": "false"
+    },
+    {
+      "Label": "DeployLoggings",
+      "Value": "true"
+    },
+    {
+      "Label": "DeployBastion",
+      "Value": "false"
+    },
+    {
+      "Label": "EnableEfsBackups",
+      "Value": "false"
+    },
+    {
+      "Label": "DeployVpc",
+      "Value": "false"
+    },
+    {
+      "Label": "ProjectName",
+      "Value": "ast-configurator"
+    },
+    {
+      "Label": "FQDN",
+      "Value": "ast-demo.cld.jpl.nasa.gov"
+    },
+    {
+      "Label": "QSS3BucketName",
+      "Value": "a2c-qs-assets"
+    },
+    {
+      "Label": "QSS3BucketPrefix",
+      "Value": "ast-aerie/"
+    },
+    {
+      "Label": "QSS3BucketRegion",
+      "Value": "us-gov-west-1"
+    },
+    {
+      "Label": "AITInstanceType",
+      "Value": "t3.medium"
+    },
+    {
+      "Label": "CloudWatchLogsRetentionPeriod",
+      "Value": "30"
+    },
+    {
+      "Label": "KeyPairName",
+      "Value": "ammos-cloud"
+    },
+    {
+      "Label": "HostedZoneID",
+      "Value": ""
+    },
+    {
+      "Label": "AccessSgID",
+      "Value": "sg-0cb679928fc3ebf29"
+    },
+    {
+      "Label": "VpcId",
+      "Value": "vpc-03456b468daa53da4"
+    },
+    {
+      "Label": "PrivateSubnet1CIDR",
+      "Value": "10.0.128.0/24"
+    },
+    {
+      "Label": "PrivateSubnet2CIDR",
+      "Value": "10.0.192.0/24"
+    },
+    {
+      "Label": "PrivateSubnet3CIDR",
+      "Value": "10.0.200.0/24"
+    },
+    { "Label": "PublicSubnet1CIDR", "Value": "10.0.32.0/24" },
+    { "Label": "PublicSubnet2CIDR", "Value": "10.0.64.0/24" },
+    { "Label": "PublicSubnet3CIDR", "Value": "10.0.70.0/24" },
+    {
+      "Label": "RemoteAccessCIDR",
+      "Value": "128.149.0.0/16"
+    }
+  ]
+}
+
+
+TemplateURL: !Sub
+- 'https://${S3Bucket}.s3.${S3Region}.${AWS::URLSuffix}/${QSS3KeyPrefix}templates/ast-iam-roles.template.yaml'
+- S3Region: !If [UsingDefaultBucket, !Ref 'AWS::Region', !Ref QSS3BucketRegion]
+  S3Bucket: !If [UsingDefaultBucket, !Sub '${QSS3BucketName}-${AWS::Region}', !Ref QSS3BucketName]
